[move/#42] config 파일 내 폴더 이동
