+++ src/main/java/org/chefcrew/config/WebConfig.java
@@ -1,11 +1,13 @@
 package org.chefcrew.config;
 
-import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.chefcrew.jwt.UserIdResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import java.util.List;
 
 @RequiredArgsConstructor
 @Configuration
