+++ src/main/java/org/chefcrew/jwt/UserId.java
@@ -1,4 +1,4 @@
-package org.chefcrew.config;
+package org.chefcrew.jwt;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
